requirements.txt
# Core dependencies
flask==2.3.3
gunicorn==21.2.0
python-dotenv==1.0.0
requests==2.31.0

# Trading and market data (FREE)
ccxt==4.1.77

# Data processing and analysis
pandas==1.5.3
numpy==1.24.3
ta==0.10.2

# Sentiment analysis (FREE)
feedparser==6.0.10
textblob==0.17.1

# Async and networking
aiohttp==3.8.6
asyncio==3.4.3
websockets==11.0.3

# Testing
pytest==7.4.2
pytest-asyncio==0.21.1

# Statistical analysis
scikit-learn==1.3.0
scipy==1.11.1

# Note: All paid API dependencies removed
# - Coinglass API replaced with CCXT order book analysis
# - TradingView API replaced with CCXT OHLCV data
# - CryptoPanic API replaced with RSS feed parsing
